feat: update swc plugin template
