+++ crates/swc_plugin_testing/src/lib.rs
@@ -1,1 +1,1 @@
-
+mod template;
